refactor(codegen): extract voidElements into plugin folder
