+++ scripts/codegen/src/semantic-checks.ts
@@ -19,6 +19,7 @@ import {
   checkTokenNames,
 } from "./plugins/tokens/check.ts";
 import { checkExamplesReferences, checkVariantChoiceKeys } from "./plugins/variants/check.ts";
+import { checkVoidElementConstraints } from "./plugins/voidElements/check.ts";
 import type { PayloadEntry, Spec, SpecPart } from "./schema.ts";
 
 type TokensCss = ReadonlySet<string>;
@@ -758,67 +759,6 @@ export function checkImperativeProps(spec: Spec, vocabulary: Vocabulary): Issue[
       );
     }
   }
-  return issues;
-}
-
-// ── Void HTML elements reject child-bearing prop declarations ───────────────
-
-/**
- * Void elements (`hr`, `img`, `input`, `br`, …) cannot have children, so
- * codegen emits the self-closing form. A spec author who declares slot props,
- * `loading`, or `as` on a void element gets no compile-time signal — the
- * codegen silently ignores them. Flag at spec-validation time so the dead
- * fields surface in review. `disabled` is rejected on non-form-control voids
- * (where it is a no-op) but accepted on `input` (where it carries native
- * semantics). Walks composite parts.
- */
-const FORM_CONTROL_VOIDS = new Set(["input"]);
-
-export function checkVoidElementConstraints(spec: Spec): Issue[] {
-  const issues: Issue[] = [];
-  visitNodes(spec, (node, path) => {
-    if (!node.element || !isVoidElement(node.element)) return;
-    const tag = node.element;
-    // `isVoidElement` lowercases its input; compare the FORM_CONTROL_VOIDS
-    // membership the same way so `element: INPUT` is treated as `element: input`.
-    const tagLower = tag.toLowerCase();
-    const propsPath = (key: string) => (path === "" ? `props.${key}` : `${path}.props.${key}`);
-    const slotProps = Object.entries(node.props ?? {})
-      .filter(([, d]) => d.slot === true)
-      .map(([n]) => n);
-    for (const slot of slotProps) {
-      issues.push(
-        issue(spec.name, propsPath(slot), `void element <${tag}> cannot host slot props`),
-      );
-    }
-    if ("loading" in (node.props ?? {})) {
-      issues.push(
-        issue(
-          spec.name,
-          propsPath("loading"),
-          `void element <${tag}> cannot render a loading spinner child`,
-        ),
-      );
-    }
-    if ("as" in (node.props ?? {})) {
-      issues.push(
-        issue(
-          spec.name,
-          propsPath("as"),
-          `void element <${tag}> cannot declare \`as\` — polymorphism to a non-void element breaks codegen-time void detection`,
-        ),
-      );
-    }
-    if ("disabled" in (node.props ?? {}) && !FORM_CONTROL_VOIDS.has(tagLower)) {
-      issues.push(
-        issue(
-          spec.name,
-          propsPath("disabled"),
-          `void element <${tag}> ignores \`disabled\` (only form-control voids like <input> honor it)`,
-        ),
-      );
-    }
-  });
   return issues;
 }
 
@@ -2315,5 +2255,6 @@ export {
   checkExamplesReferences,
   checkVariantChoiceKeys,
 } from "./plugins/variants/check.ts";
+export { checkVoidElementConstraints } from "./plugins/voidElements/check.ts";
 
 export type { CssIndex };
